Add missing indendations in comments.

+++ components/script/dom/xrinputsource.rs
@@ -108,7 +108,7 @@ impl XRInputSource {
             });
         frame.axis_values.iter().enumerate().for_each(|(i, value)| {
             self.gamepad
-                .map_and_normalize_axes(i as usize, *value as f64);
+                .map_and_normalize_axes(i, *value as f64);
         });
     }
 }
